+++ database/java/migrations/ticket-order-db-migrations/ticket-order-analytical-migrations/src/main/resources/db/repeatable/R__enabled_app_users.sql
@@ -1,0 +1,9 @@
+CREATE OR REPLACE VIEW ticket_analytical.enabled_app_users AS
+SELECT
+    id,
+    email,
+    role,
+    created_at,
+    updated_at
+FROM ticket_transactional.t_app_users
+WHERE enabled = TRUE;