+++ database/java/migrations/ticket-order-db-migrations/ticket-order-transactional-migrations/src/main/resources/db/repeatable/R__app_users.sql
@@ -1,0 +1,9 @@
+CREATE OR REPLACE VIEW ticket_transactional.app_users AS
+SELECT
+    id,
+    email,
+    role,
+    enabled,
+    created_at,
+    updated_at
+FROM ticket_transactional.t_app_users;